[FIX] spring security config not to user webSecurityConfiguer
