--- server/user-service/src/main/java/org/wid/userservice/config/SecurityConfig.java
@@ -1,36 +1,0 @@
-package org.wid.userservice.config;
-
-import org.springframework.boot.autoconfigure.security.reactive.PathRequest;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-import org.wid.userservice.port.primary.AuthServicePort;
-
-import lombok.RequiredArgsConstructor;
-
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-  private final AuthServicePort authService;
-
-  @Bean
-  public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-    http.csrf(AbstractHttpConfigurer::disable);
-    http.httpBasic(AbstractHttpConfigurer::disable);
-
-    return http.build();
-  }
-
-  @Bean
-  public WebSecurityCustomizer webSecurityCustomizer() {
-    return web -> web
-        .ignoring()
-        .requestMatchers(String.valueOf(PathRequest.toStaticResources().atCommonLocations()));
-  }
-}
